+++ crates/bedit-cloudburst/src/lib.rs
@@ -1,13 +1,17 @@
 //! `bedit-cloudburst` provides strongly typed data structures for serializing and deserializing torrents.
 
 mod fileattributes;
+mod crypto;
 mod files;
 mod info;
+mod pieces;
 mod signature;
 mod torrent;
 
 pub use fileattributes::{FileAttribute, TorrentFileAttributes};
+pub use crypto::Sha1Hash;
 pub use files::{FileTree, FileTreeEntry, FileTreeInfo, SharedFiles};
-pub use info::{Hybrid, Info, MetaV1, MetaV2, PieceLength};
+pub use info::{Hybrid, Info, MetaV1, MetaV2};
+pub use pieces::{PieceLength, Pieces};
 pub use signature::{SignInfo, Signature};
 pub use torrent::{Node, Torrent};